Put futures macros into a seperate mod
So it looks cleaner.
